requirements.txt
python-telegram-bot==21.4
requests==2.32.3
fastapi==0.112.0
uvicorn